build,win: goto lint only after defining node_exe
When running `vcbuild lint` on a new prompt where vcbuild was not run
before, `%node_exe%` was not defined and `eslint.js`, `lint-js.js`
and `lint-md.js` would be run directly using the program defined in
Windows (usually the Windows Scripting Host or Notepad).

This moves the goto statement to after `%node_exe%` is defined.

Fixes: https://github.com/nodejs/node/issues/29602

PR-URL: https://github.com/nodejs/node/pull/29616
Reviewed-By: Ruben Bridgewater <ruben@bridgewater.de>