doc: de-duplicate security release processes
The security release process is spread across multiple files. Merge
these two files to remove duplication and inconsistency. Also, make the
format more useful for inserting into the description of the Next
Security Release issue description.

This seems an obvious candidate for a github issue template, but if it
was, the content would not be reviewable by anyone outside of those on
the security teams, and the process should be public for purposes of
transparency and review.

PR-URL: https://github.com/nodejs/node/pull/30996
Reviewed-By: Rich Trott <rtrott@gmail.com>